refactor(services): playground_service 를 패키지로 승격한다 — 2 단계 (B5.5 Task 5)
`services/playground_service.py` 1,286 줄 → 패키지 5 모듈, 최대 738 줄.
B5.5 마지막이자 **유일한 2 단계**다.

    config.py    76줄  플래그 · 경로 · 기본 프롬프트 · 목 도구 목록
    storage.py  275줄  파일/DB 영속화 + 인메모리 캐시
    llm.py      179줄  LangChain 변환 · 모델 폴백
    mock.py      91줄  목 응답 생성기
    service.py  738줄  PlaygroundService (13 메서드)

**1 단계 — 메서드 추출.** `PlaygroundService` 가 혼자 808 줄이라 정의 이동만으로는
못 들어간다. DB 그룹 5 개를 클래스 밖으로 들어올렸다(808 → 687). 클래스 전체가
`@staticmethod` 라 네임스페이스이므로 모듈 함수로 옮긴 뒤
`name = staticmethod(name)` 으로 재부착하면 `PlaygroundService.save_session_to_db(...)`
호출 형태가 완전히 보존된다(mypy 통과 확인 — 위임 함수 25 줄이 불필요해졌다).

DB 그룹을 고른 것은 크기가 아니라 **제약 두 개를 한 번에 풀기 때문이다.**
`_initialized` 재바인딩이 `_load_sessions` 와 `load_sessions_from_db` 둘인데 후자가
클래스 안이라 48 + 808 = 856 줄이 한 모듈로 묶여 있었다. 들어올리니 둘이 `storage`
로 모이고 클래스도 줄었다.

도구로 남겼다 — `tests/backend/api/lift_staticmethods.py`. B6 8 개가 같은 성질이다.
`@staticmethod` 가 아니면 중단하고, 들어올린 뒤에도 클래스를 참조하면(순환 import 가
된다) 중단한다. 실측: 한정 참조는 정확히 2 건이었고 그것만 해제됐다.

**AST 로 재바인딩과 변형을 구분한 것이 배정표의 근거다.**
`global _sessions, _initialized` 가 두 곳에 있지만 `_sessions` 는 첨자 대입·삭제만
되고 **한 번도 재바인딩되지 않는다**(그 `global` 선언은 잉여다). 그래서 여러 모듈이
같은 dict 를 보고 배럴 재노출도 안전하다. 진짜 제약은 `_initialized` 하나였다.

**패치 12 건은 앞의 세 태스크와 방향이 반대다.**
모듈 **객체** 패치이고, 세 이름(`_load_sessions`·`_save_sessions`·`_fire_and_forget`)의
정의는 `storage` 로 가지만 **호출 지점 21 곳이 전부 `service`** 다(실측: storage 범위
0 건). `from .storage import ...` 는 바인딩을 복사하므로 패치는 정의처가 아니라
읽는 쪽 — `playground_service.service` — 을 겨냥해야 한다. RED 확인 후 49 통과.

**`STORAGE_DIR` 은 #358 의 `DATA_DIR` 과 같은 깊이 결합 경로다.** 본문을 고쳤고
Red-Green 으로 확인했다(원래 깊이로 되돌리니 그 단언 하나만 실패).

**감사를 두 단계로 나눴다.** 1 단계 산출물 대비는 정의 28 종 전부 순수 이동으로
통과하고, 원본 대비는 1 단계 변경분(신규 5 종 + 클래스 본문 1 건)만 정확히 드러난다.
한 번에 감사했으면 이동과 편집을 구분할 수 없었다.

**검증** — 분할 전후 실행 비교(STORAGE_DIR · _DB_AVAILABLE · 재부착 5 종 호출 가능성 ·
공유 dict · 메서드 표면 · 순수 함수) JSON diff 완전 일치, 계약 테스트 19 건에
Red-Green 3 종, ruff · ruff format · mypy(331) · pytest 1,754 passed / 0 failed.
