+++ src/backend/services/playground_service/mock.py
@@ -1,0 +1,91 @@
+"""LLM 없이 동작을 확인할 때 쓰는 목 응답 생성기."""
+
+from typing import Any
+
+
+def _generate_mock_response(
+    prompt: str,
+    agent_id: str | None,
+    tools: list[str],
+) -> str:
+    """Generate a mock LLM response for testing."""
+    agent_name = agent_id or "default agent"
+
+    if "code" in prompt.lower() or "function" in prompt.lower():
+        return f"""I'll help you with that coding task. Here's a solution:
+
+```python
+def example_function():
+    # Implementation based on your request
+    result = process_data()
+    return result
+```
+
+This implementation follows best practices and handles edge cases. Let me know if you need any modifications!
+
+[Generated by {agent_name} in playground mode]"""
+
+    elif "explain" in prompt.lower() or "what is" in prompt.lower():
+        return f"""Great question! Let me explain:
+
+The concept you're asking about involves several key components:
+
+1. **First Component**: This handles the initial processing
+2. **Second Component**: This manages the core logic
+3. **Third Component**: This produces the final output
+
+Each component works together to achieve the desired result. Would you like me to elaborate on any specific part?
+
+[Generated by {agent_name} in playground mode]"""
+
+    else:
+        return f"""I understand your request. Here's my response:
+
+Based on the context provided, I would recommend the following approach:
+
+- Start by analyzing the requirements
+- Break down the task into smaller steps
+- Implement each step carefully
+- Test and validate the results
+
+Available tools for this session: {", ".join(tools) if tools else "None"}
+
+Is there anything specific you'd like me to focus on?
+
+[Generated by {agent_name} in playground mode]"""
+
+
+def _generate_mock_tool_result(tool_name: str, arguments: dict[str, Any]) -> Any:
+    """Generate mock tool execution result."""
+    if tool_name == "web_search":
+        return {
+            "results": [
+                {"title": "Result 1", "url": "https://example.com/1", "snippet": "..."},
+                {"title": "Result 2", "url": "https://example.com/2", "snippet": "..."},
+            ],
+            "total": 2,
+        }
+    elif tool_name == "code_execute":
+        return {
+            "output": "Hello, World!",
+            "exit_code": 0,
+            "execution_time_ms": 50,
+        }
+    elif tool_name == "file_read":
+        return {
+            "content": "# Sample file content\nThis is a test file.",
+            "size": 42,
+        }
+    elif tool_name == "file_write":
+        return {
+            "success": True,
+            "path": arguments.get("path", "unknown"),
+            "bytes_written": len(arguments.get("content", "")),
+        }
+    elif tool_name == "api_call":
+        return {
+            "status": 200,
+            "body": {"message": "Success"},
+            "headers": {"content-type": "application/json"},
+        }
+    return {"result": "Mock result"}
