+++ src/backend/services/playground_service/service.py
@@ -1,20 +1,21 @@
-"""Playground service for agent testing environment."""
+"""PlaygroundService — 세션 CRUD · 실행 · 도구 · 비교.
 
-import asyncio
+DB 영속화 5 함수는 `storage.py` 로 들어올린 뒤 클래스에 같은 이름으로
+재부착했다(`staticmethod(...)`) — `PlaygroundService.save_session_to_db(...)`
+호출 형태가 그대로 유지된다. 클래스 전체가 `@staticmethod` 라 가능한 형태다.
+
+`_load_sessions` · `_save_sessions` · `_fire_and_forget` 을 읽는 곳이 전부
+이 모듈이다. 테스트는 이 모듈 객체를 패치한다 —
+`monkeypatch.setattr(playground_service.service, "_load_sessions", ...)`.
+"""
+
 import json
 import logging
-import os
 from collections.abc import AsyncIterator
-from datetime import datetime
-from pathlib import Path
 from typing import Any
-
-from langchain_core.messages import AIMessage, BaseMessage, HumanMessage, SystemMessage
 
 from models.cost import calculate_cost, estimate_tokens
 from models.llm_access import LLMAccessResponse
-from models.llm_models import LLMModelRegistry
-from models.llm_usage import LLMUsageSource
 from models.playground import (
     PlaygroundCompareRequest,
     PlaygroundCompareResult,
@@ -26,358 +27,31 @@ from models.playground import (
     PlaygroundSessionCreate,
     PlaygroundToolTest,
 )
-from services.llm_runtime_resolver import (
-    LLMRuntimeRequest,
-    LLMRuntimeResolutionError,
-    resolve_llm_runtime,
-)
-from services.llm_service import LLMResponse, LLMService
+from services.llm_service import LLMService
 from services.playground_context import build_effective_system_prompt
 from utils.time import utcnow
 
+from .config import DEFAULT_SYSTEM_PROMPT, PLAYGROUND_TOOLS
+from .llm import (
+    _coerce_llm_content,
+    _invoke_with_model_fallback,
+    _playground_usage_context,
+    _to_lc_messages,
+)
+from .mock import _generate_mock_tool_result
+from .storage import (
+    _fire_and_forget,
+    _load_sessions,
+    _model_to_pydantic,
+    _pydantic_to_db_dict,
+    _save_sessions,
+    _sessions,
+    delete_session_from_db,
+    load_sessions_from_db,
+    save_session_to_db,
+)
+
 logger = logging.getLogger(__name__)
-
-# ─────────────────────────────────────────────────────────────
-# Database Toggle
-# ─────────────────────────────────────────────────────────────
-
-USE_DATABASE = os.getenv("USE_DATABASE", "false").lower() == "true"
-
-# Conditional DB imports (only when USE_DATABASE=true)
-try:
-    from sqlalchemy import select
-    from sqlalchemy.ext.asyncio import AsyncSession
-
-    from db.models.playground import PlaygroundSessionModel
-
-    _DB_AVAILABLE = True
-except ImportError:
-    _DB_AVAILABLE = False
-    AsyncSession = Any  # type: ignore
-    PlaygroundSessionModel = None  # type: ignore
-
-# Persistent storage file path
-STORAGE_DIR = Path(__file__).parent.parent / "data"
-SESSIONS_FILE = STORAGE_DIR / "playground_sessions.json"
-
-# In-memory cache
-_sessions: dict[str, PlaygroundSession] = {}
-_initialized = False
-
-# Default system prompt for playground — Gemini 특화 가이드
-DEFAULT_SYSTEM_PROMPT = """당신은 AOS 플랫폼의 AI 어시스턴트입니다.
-
-- 한국어로 답변하고, 코드/명령어/경로는 원문을 그대로 유지합니다.
-- project context가 제공되면 그것을 사실의 단일 출처로 삼고,
-  컨텍스트에 없는 사실은 추측 대신 "제공된 컨텍스트에 없습니다"라고 명시합니다.
-- 답하기 전에 의도와 필요한 정보를 1~2문장으로 정리한 뒤 답변합니다.
-- 외부/최신 정보가 필요하면 사용 가능한 도구를 우선 호출합니다.
-- 여러 단계가 필요한 작업은 단계별로 명확히 설명합니다."""
-
-
-def _to_lc_messages(msgs: list[PlaygroundMessage]) -> list[BaseMessage]:
-    """
-    Convert persisted ``PlaygroundMessage`` history into LangChain messages.
-
-    - ``user`` → ``HumanMessage``
-    - ``assistant`` → ``AIMessage``
-    - ``tool`` → ``SystemMessage(content="[previous tool call] ...")``
-      (Legacy sessions don't carry ``tool_call_id`` required by LangChain
-      ``ToolMessage``; absorbing as SystemMessage preserves the information
-      without tripping strict tool-call validation on the next turn.)
-    - ``system`` role is omitted (handled via explicit ``system_prompt`` arg)
-    """
-    out: list[BaseMessage] = []
-    for m in msgs:
-        role = m.role
-        content = m.content or ""
-        if role == "user":
-            out.append(HumanMessage(content=content))
-        elif role == "assistant":
-            out.append(AIMessage(content=content))
-        elif role == "tool":
-            out.append(SystemMessage(content=f"[previous tool call] {content}"))
-        # other roles (system, etc.) are intentionally dropped
-    return out
-
-
-def _coerce_llm_content(content: Any) -> str:
-    """
-    Flatten LLM content into a plain string.
-
-    Gemini's multi-part responses can arrive as ``[{type: "text", text: ...},
-    {type: "tool_use", ...}]``. We extract only the textual parts so the
-    ``tool_use`` blocks don't leak into persisted transcripts as ``str(dict)``.
-    """
-    if isinstance(content, str):
-        return content
-    if isinstance(content, list):
-        parts: list[str] = []
-        for item in content:
-            if isinstance(item, dict):
-                # Only collect text parts; skip tool_use/tool_result blocks
-                if item.get("type") in (None, "text"):
-                    text = item.get("text")
-                    if text:
-                        parts.append(str(text))
-            else:
-                parts.append(str(item))
-        return " ".join(parts)
-    return str(content)
-
-
-def _playground_usage_context(
-    session: PlaygroundSession,
-    *,
-    execution: PlaygroundExecution | None = None,
-    metadata: dict[str, Any] | None = None,
-    llm_access: LLMAccessResponse | None = None,
-) -> dict[str, Any]:
-    context_metadata = {
-        "agent_id": session.agent_id,
-        "execution_id": execution.id if execution else None,
-    }
-    if metadata:
-        context_metadata.update(metadata)
-    user_id = session.user_id or (llm_access.user_id if llm_access else None)
-    context = {
-        "source": LLMUsageSource.PLAYGROUND,
-        "session_id": session.id,
-        "project_id": session.project_id,
-        "metadata": {k: v for k, v in context_metadata.items() if v is not None},
-    }
-    if user_id:
-        context["user_id"] = user_id
-    if llm_access:
-        context["llm_access"] = llm_access
-    return context
-
-
-def _is_inaccessible_model_error(error: Exception) -> bool:
-    """Detect errors caused by a model the current caller cannot use.
-
-    Two distinct sources must trigger the stale-model fallback:
-    - provider-side errors (OpenAI/Anthropic "model_not_found", etc.), matched by
-      message, and
-    - the runtime resolver rejecting the requested model for the authenticated
-      user's entitlements. ``LLMService.invoke`` re-raises that
-      ``LLMRuntimeResolutionError`` raw, and its message ("No enabled LLM
-      entitlement for provider=...") matches none of the provider patterns — so
-      it must be recognized by type, or authenticated fallback never fires.
-    """
-    if isinstance(error, LLMRuntimeResolutionError):
-        return True
-    message = str(error).lower()
-    return (
-        "model_not_found" in message
-        or "does not have access to model" in message
-        or "unknown model:" in message
-    )
-
-
-def _safe_playground_fallback_model(
-    current_model: str,
-    usage_context: dict[str, Any] | None = None,
-) -> str | None:
-    """Pick a conservative fallback for stale persisted Playground sessions.
-
-    For an authenticated user the runtime resolver gates models by entitlement,
-    so the fallback must be a model the user is actually entitled to. The global
-    env default may not be in their entitlements and would only trigger a fresh
-    ``LLMRuntimeResolutionError`` — breaking the fallback. Resolve the user's
-    entitled default instead; return ``None`` (no doomed retry) when they have no
-    usable entitlement. Anonymous sessions keep the env-default behaviour.
-    """
-    access = usage_context.get("llm_access") if usage_context else None
-    if isinstance(access, LLMAccessResponse):
-        try:
-            resolution = resolve_llm_runtime(
-                access,
-                LLMRuntimeRequest(
-                    user_id=usage_context.get("user_id") if usage_context else None,
-                    source=LLMUsageSource.PLAYGROUND.value,
-                    requested_model_id=None,
-                    organization_id=(
-                        usage_context.get("organization_id") if usage_context else None
-                    ),
-                ),
-            )
-        except LLMRuntimeResolutionError:
-            return None
-        entitled = resolution.model_id
-        return entitled if entitled and entitled != current_model else None
-
-    configured_default = LLMModelRegistry.get_default(os.getenv("LLM_PROVIDER") or "codex_cli")
-    fallback = configured_default or "codex-cli"
-    return fallback if fallback and fallback != current_model else None
-
-
-async def _invoke_with_model_fallback(
-    session: PlaygroundSession,
-    invoke,
-    **kwargs: Any,
-) -> LLMResponse:
-    """Invoke once, then retry with the configured safe default for stale models."""
-    try:
-        return await invoke(model_id=session.model, **kwargs)
-    except Exception as exc:
-        fallback_model = _safe_playground_fallback_model(session.model, kwargs.get("usage_context"))
-        if not fallback_model or not _is_inaccessible_model_error(exc):
-            raise
-
-        stale_model = session.model
-        logger.warning(
-            "playground_model_inaccessible_retry",
-            extra={"stale_model": stale_model, "fallback_model": fallback_model},
-        )
-        session.model = fallback_model
-        return await invoke(model_id=fallback_model, **kwargs)
-
-
-def _ensure_storage_dir():
-    """Ensure storage directory exists."""
-    STORAGE_DIR.mkdir(parents=True, exist_ok=True)
-
-
-def _load_sessions():
-    """Load sessions from persistent storage."""
-    global _sessions, _initialized
-    if _initialized:
-        return
-
-    _ensure_storage_dir()
-
-    if SESSIONS_FILE.exists():
-        try:
-            with open(SESSIONS_FILE, encoding="utf-8") as f:
-                data = json.load(f)
-                for session_data in data:
-                    # Convert datetime strings back to datetime objects
-                    if "created_at" in session_data and isinstance(session_data["created_at"], str):
-                        session_data["created_at"] = datetime.fromisoformat(
-                            session_data["created_at"]
-                        )
-                    if "updated_at" in session_data and isinstance(session_data["updated_at"], str):
-                        session_data["updated_at"] = datetime.fromisoformat(
-                            session_data["updated_at"]
-                        )
-
-                    # Convert messages
-                    if "messages" in session_data:
-                        for msg in session_data["messages"]:
-                            if "timestamp" in msg and isinstance(msg["timestamp"], str):
-                                msg["timestamp"] = datetime.fromisoformat(msg["timestamp"])
-
-                    # Convert executions
-                    if "executions" in session_data:
-                        for exec_data in session_data["executions"]:
-                            for dt_field in ["created_at", "started_at", "completed_at"]:
-                                if dt_field in exec_data and isinstance(exec_data[dt_field], str):
-                                    exec_data[dt_field] = datetime.fromisoformat(
-                                        exec_data[dt_field]
-                                    )
-                            if "messages" in exec_data:
-                                for msg in exec_data["messages"]:
-                                    if "timestamp" in msg and isinstance(msg["timestamp"], str):
-                                        msg["timestamp"] = datetime.fromisoformat(msg["timestamp"])
-
-                    session = PlaygroundSession(**session_data)
-                    _sessions[session.id] = session
-        except Exception as e:
-            print(f"Warning: Failed to load playground sessions: {e}")
-
-    _initialized = True
-
-
-def _save_sessions():
-    """Save sessions to persistent storage."""
-    _ensure_storage_dir()
-
-    try:
-        data = []
-        for session in _sessions.values():
-            session_dict = session.model_dump()
-            # Convert datetime objects to ISO strings for JSON serialization
-            if session_dict.get("created_at"):
-                session_dict["created_at"] = session_dict["created_at"].isoformat()
-            if session_dict.get("updated_at"):
-                session_dict["updated_at"] = session_dict["updated_at"].isoformat()
-
-            # Convert messages
-            for msg in session_dict.get("messages", []):
-                if msg.get("timestamp"):
-                    msg["timestamp"] = msg["timestamp"].isoformat()
-
-            # Convert executions
-            for exec_data in session_dict.get("executions", []):
-                for dt_field in ["created_at", "started_at", "completed_at"]:
-                    if exec_data.get(dt_field):
-                        exec_data[dt_field] = exec_data[dt_field].isoformat()
-                for msg in exec_data.get("messages", []):
-                    if msg.get("timestamp"):
-                        msg["timestamp"] = msg["timestamp"].isoformat()
-
-            data.append(session_dict)
-
-        with open(SESSIONS_FILE, "w", encoding="utf-8") as f:
-            json.dump(data, f, ensure_ascii=False, indent=2)
-    except Exception as e:
-        print(f"Warning: Failed to save playground sessions: {e}")
-
-
-# Mock tools for playground
-PLAYGROUND_TOOLS = {
-    "web_search": {
-        "name": "web_search",
-        "description": "Search the web for information",
-        "parameters": {
-            "query": {"type": "string", "description": "Search query"},
-            "max_results": {"type": "integer", "description": "Max results", "default": 5},
-        },
-    },
-    "code_execute": {
-        "name": "code_execute",
-        "description": "Execute code in a sandbox",
-        "parameters": {
-            "language": {"type": "string", "description": "Programming language"},
-            "code": {"type": "string", "description": "Code to execute"},
-        },
-    },
-    "file_read": {
-        "name": "file_read",
-        "description": "Read a file from the workspace",
-        "parameters": {
-            "path": {"type": "string", "description": "File path"},
-        },
-    },
-    "file_write": {
-        "name": "file_write",
-        "description": "Write content to a file",
-        "parameters": {
-            "path": {"type": "string", "description": "File path"},
-            "content": {"type": "string", "description": "Content to write"},
-        },
-    },
-    "api_call": {
-        "name": "api_call",
-        "description": "Make an HTTP API call",
-        "parameters": {
-            "method": {"type": "string", "description": "HTTP method"},
-            "url": {"type": "string", "description": "URL to call"},
-            "body": {"type": "object", "description": "Request body"},
-        },
-    },
-}
-
-
-def _fire_and_forget(coro) -> None:
-    """Schedule a coroutine as fire-and-forget if an event loop is running."""
-    try:
-        loop = asyncio.get_event_loop()
-        if loop.is_running():
-            asyncio.create_task(coro)
-    except RuntimeError:
-        pass
 
 
 class PlaygroundService:
@@ -1055,232 +729,10 @@ class PlaygroundService:
     # Async Database Methods
     # ─────────────────────────────────────────────────────────────
 
-    @staticmethod
-    def _model_to_pydantic(model: "PlaygroundSessionModel") -> PlaygroundSession:
-        """Convert SQLAlchemy model to Pydantic model."""
-        return PlaygroundSession(
-            id=model.id,
-            name=model.name,
-            description=model.description or "",
-            user_id=model.user_id,
-            project_id=model.project_id,
-            working_directory=model.working_directory,
-            agent_id=model.agent_id,
-            model=model.model or "",
-            temperature=model.temperature or 0.7,
-            max_tokens=model.max_tokens or 4096,
-            system_prompt=model.system_prompt,
-            rag_enabled=model.rag_enabled or False,
-            rag_k=getattr(model, "rag_k", None) or 5,
-            rag_hybrid_override=getattr(model, "rag_hybrid_override", None),
-            rag_rerank_override=getattr(model, "rag_rerank_override", None),
-            rag_include_shared=bool(getattr(model, "rag_include_shared", False)),
-            rules_mode=getattr(model, "rules_mode", None) or "off",
-            memory_mode=getattr(model, "memory_mode", None) or "off",
-            selected_rule_ids=list(getattr(model, "selected_rule_ids", None) or []),
-            selected_memory_ids=list(getattr(model, "selected_memory_ids", None) or []),
-            context_budget_tokens=getattr(model, "context_budget_tokens", None) or 8000,
-            available_tools=model.available_tools or [],
-            enabled_tools=model.enabled_tools or [],
-            messages=[PlaygroundMessage(**m) for m in (model.messages or [])],
-            executions=[PlaygroundExecution(**e) for e in (model.executions or [])],
-            total_executions=model.total_executions or 0,
-            total_tokens=model.total_tokens or 0,
-            total_cost=model.total_cost or 0.0,
-            created_at=model.created_at,
-            updated_at=model.updated_at,
-        )
-
-    @staticmethod
-    def _pydantic_to_db_dict(session: PlaygroundSession) -> dict:
-        """Convert Pydantic model to dict for DB upsert."""
-        return {
-            "name": session.name,
-            "description": session.description,
-            "user_id": session.user_id,
-            "project_id": session.project_id,
-            "working_directory": session.working_directory,
-            "agent_id": session.agent_id,
-            "model": session.model,
-            "temperature": session.temperature,
-            "max_tokens": session.max_tokens,
-            "system_prompt": session.system_prompt,
-            "rag_enabled": session.rag_enabled,
-            "rag_k": session.rag_k,
-            "rag_hybrid_override": session.rag_hybrid_override,
-            "rag_rerank_override": session.rag_rerank_override,
-            "rag_include_shared": session.rag_include_shared,
-            "rules_mode": session.rules_mode,
-            "memory_mode": session.memory_mode,
-            "selected_rule_ids": list(session.selected_rule_ids),
-            "selected_memory_ids": list(session.selected_memory_ids),
-            "context_budget_tokens": session.context_budget_tokens,
-            "available_tools": session.available_tools,
-            "enabled_tools": session.enabled_tools,
-            "messages": [m.model_dump(mode="json") for m in session.messages],
-            "executions": [e.model_dump(mode="json") for e in session.executions],
-            "total_executions": session.total_executions,
-            "total_tokens": session.total_tokens,
-            "total_cost": session.total_cost,
-            "updated_at": utcnow(),
-        }
-
-    @staticmethod
-    async def save_session_to_db(session: PlaygroundSession) -> None:
-        """Save/update a session to the database."""
-        if not USE_DATABASE or not _DB_AVAILABLE:
-            return
-        try:
-            from db.database import async_session_factory
-
-            async with async_session_factory() as db:
-                result = await db.execute(
-                    select(PlaygroundSessionModel).where(PlaygroundSessionModel.id == session.id)
-                )
-                existing = result.scalar_one_or_none()
-                data = PlaygroundService._pydantic_to_db_dict(session)
-                if existing:
-                    for key, value in data.items():
-                        setattr(existing, key, value)
-                else:
-                    model = PlaygroundSessionModel(
-                        id=session.id, **data, created_at=session.created_at
-                    )
-                    db.add(model)
-                await db.commit()
-        except Exception:
-            logger.warning("Failed to save playground session %s to DB", session.id)
-
-    @staticmethod
-    async def delete_session_from_db(session_id: str) -> None:
-        """Delete a session from the database."""
-        if not USE_DATABASE or not _DB_AVAILABLE:
-            return
-        try:
-            from db.database import async_session_factory
-
-            async with async_session_factory() as db:
-                result = await db.execute(
-                    select(PlaygroundSessionModel).where(PlaygroundSessionModel.id == session_id)
-                )
-                existing = result.scalar_one_or_none()
-                if existing:
-                    await db.delete(existing)
-                    await db.commit()
-        except Exception:
-            logger.warning("Failed to delete playground session %s from DB", session_id)
-
-    @staticmethod
-    async def load_sessions_from_db() -> None:
-        """Load sessions from DB into in-memory cache (startup sync)."""
-        if not USE_DATABASE or not _DB_AVAILABLE:
-            return
-        try:
-            from db.database import async_session_factory
-
-            global _sessions, _initialized
-            async with async_session_factory() as db:
-                result = await db.execute(select(PlaygroundSessionModel))
-                models = result.scalars().all()
-                for model in models:
-                    if model.id not in _sessions:
-                        _sessions[model.id] = PlaygroundService._model_to_pydantic(model)
-                _initialized = True
-        except Exception:
-            logger.warning("Failed to load playground sessions from DB")
-
-
-# ─────────────────────────────────────────────────────────────
-# Helper Functions
-# ─────────────────────────────────────────────────────────────
-
-
-def _generate_mock_response(
-    prompt: str,
-    agent_id: str | None,
-    tools: list[str],
-) -> str:
-    """Generate a mock LLM response for testing."""
-    agent_name = agent_id or "default agent"
-
-    if "code" in prompt.lower() or "function" in prompt.lower():
-        return f"""I'll help you with that coding task. Here's a solution:
-
-```python
-def example_function():
-    # Implementation based on your request
-    result = process_data()
-    return result
-```
-
-This implementation follows best practices and handles edge cases. Let me know if you need any modifications!
-
-[Generated by {agent_name} in playground mode]"""
-
-    elif "explain" in prompt.lower() or "what is" in prompt.lower():
-        return f"""Great question! Let me explain:
-
-The concept you're asking about involves several key components:
-
-1. **First Component**: This handles the initial processing
-2. **Second Component**: This manages the core logic
-3. **Third Component**: This produces the final output
-
-Each component works together to achieve the desired result. Would you like me to elaborate on any specific part?
-
-[Generated by {agent_name} in playground mode]"""
-
-    else:
-        return f"""I understand your request. Here's my response:
-
-Based on the context provided, I would recommend the following approach:
-
-- Start by analyzing the requirements
-- Break down the task into smaller steps
-- Implement each step carefully
-- Test and validate the results
-
-Available tools for this session: {", ".join(tools) if tools else "None"}
-
-Is there anything specific you'd like me to focus on?
-
-[Generated by {agent_name} in playground mode]"""
-
-
-def _generate_mock_tool_result(tool_name: str, arguments: dict[str, Any]) -> Any:
-    """Generate mock tool execution result."""
-    if tool_name == "web_search":
-        return {
-            "results": [
-                {"title": "Result 1", "url": "https://example.com/1", "snippet": "..."},
-                {"title": "Result 2", "url": "https://example.com/2", "snippet": "..."},
-            ],
-            "total": 2,
-        }
-    elif tool_name == "code_execute":
-        return {
-            "output": "Hello, World!",
-            "exit_code": 0,
-            "execution_time_ms": 50,
-        }
-    elif tool_name == "file_read":
-        return {
-            "content": "# Sample file content\nThis is a test file.",
-            "size": 42,
-        }
-    elif tool_name == "file_write":
-        return {
-            "success": True,
-            "path": arguments.get("path", "unknown"),
-            "bytes_written": len(arguments.get("content", "")),
-        }
-    elif tool_name == "api_call":
-        return {
-            "status": 200,
-            "body": {"message": "Success"},
-            "headers": {"content-type": "application/json"},
-        }
-    return {"result": "Mock result"}
-
-
-# Note: _calculate_cost removed - use models.cost.calculate_cost instead
+    # 모듈 레벨로 들어올린 뒤 같은 이름으로 재부착한다 —
+    # `C.f(...)` 호출 형태가 그대로 유지된다.
+    _model_to_pydantic = staticmethod(_model_to_pydantic)
+    _pydantic_to_db_dict = staticmethod(_pydantic_to_db_dict)
+    save_session_to_db = staticmethod(save_session_to_db)
+    delete_session_from_db = staticmethod(delete_session_from_db)
+    load_sessions_from_db = staticmethod(load_sessions_from_db)
